Rename to `boondock` because we're starting to diverge from upstream
We're going to go ahead and fork this crate for now.  The rationale:

- We've broken several existing APIs during our refactorings.
- The upstream maintainers haven't worked on it in almost a year, and if
  we want `cage` to be able to respond quickly to Docker-related bugs,
  then we need something that we can release ourselves.
- Docker, Inc., has asked open source projects to not begin their names
  with `docker` in order to prevent confusion.

"Boondock" was chosen as a name because I live in the boondocks of
Vermont.

Of course, the upstream `rust-docker` maintainers are welcome to use any
of the new code they wish!  Or if they'd like to contribute anything to
this fork for any reason, they're welcome to do that, too.