Support upload tar archive to create new images
Docker allows users to send a tar or tar.gz archive in the body
of a `/images/create` API call. This allows for creating completely
custom images with any file system without using a Dockerfile.

Bollard previously didn't support this API since it always sends
an empty body with requests to `/images/create` (`Docker::create_image`).
This commit adds an optional argumement to include the request body
with a new image.

There is also an example, `/examples/image_from_scratch.rs`, which shows
how to upload a custom tar file as an archive.

The commit breaks backwards compatability for the library since it adds
a new argument to the `Docker::create_image` function that must
be specified (although existing users would just need `None`).