Take all bytes from src to avoid processing again
This is necessary because the src is mutable and is a container that
gets new data appended later, if we don't update the "read" pointer of
the BytesMut we end up processing the same JSON "slice" over and over,
and the event stream never seems to progress.

This only happens on macOS where apparently the /events endpoint does
not get newline delineated.