Switch from pin-project to pin-project-lite
This is another small step towards reducing the size of bollard's
dependency tree :-)

`tokio` and `hyper` have both switched from `pin-project` crate
to the lighter-weight `pin-project-lite`:
https://github.com/tokio-rs/tokio/pull/1778
https://github.com/hyperium/hyper/pull/2566

This does the same for bollard. For the differences between the
two crates, see:
https://docs.rs/pin-project-lite/0.2.8/pin_project_lite/#pin-project-vs-pin-project-lite

Note: The full advantage of this won't be seen until a new
`hyperlocal` release exists that contains:
https://github.com/softprops/hyperlocal/pull/54

...and bollard updates to that release, so that `pin-project` can be
fully dropped from the dependency tree.